course MDL-24631 Improve get_coursemodule_info to make 'label' generic
These changes do the following:
- Add parameters ->url ->content and ->extraclasses to get_coursemodule_info api
- Add constant MOD_URL_NOLINK in course/lib.php for use with ->url parameter
- Implement these parameters in print_section in course/lib.php
- Implement parameters in all other areas of code which previously special-cased label

Note that as part of this change, the upgrade for label will call rebuild_course_cache to clear course cache for all courses. This is because label changes from using the old (weird) method of putting ->name in its modinfo, to the new method of using ->content.