MDL-44070 Quiz: Remove groupmembersonly check on override page
The quiz override page does checks related to the groupmembersonly
setting so that it can avoid showing users (or show them as not
'active') if they don't meet that requirement.

While this may be a logical thing to do, it does not fit well
with the new API. There are also other ways that users could be
configured not to access the quiz which were not taken account of.
I have discussed this with Tim and he thinks these checks are not
necessary.