MDL-44070 Workshop: Remove groupmembersonly checks
Logic relating to allocating users and showing lists of participants
used to take the 'groupmembersonly' flag into account.

I have changed the code so that it behaves as if
enablegroupmembersonly (which was always listed as an experimental
feature) had been turned off.

This will cause the following changes to behaviour:

Manual assignment:
* If groupmembersonly was enabled before, the list of potential
  users (under 'Manual allocation') would include only users who
  are in groups. Now it includes all enrolled users.

Random assignment:
* If groupmembersonly was enabled before, AND the assignment is
  set to 'no groups' mode, AND you choose the 'partipants can
  assess without having submitted anything' option, then the
  random allocation system would previously not allocate
  assignments for review to users who were not in a group. Now
  it will allocate to these users, even though (because of
  groupmembersonly) they cannot access the activity.

Note: Both before and now, it is possible for random assignment
to allocate assignment to users who cannot access the activity
for other reasons e.g. if there is a user profile condition.