MDL-44070 Conditional availability enhancements (8): global changes
This commit contains changes required to make core Moodle (and
included plugins) integrate with the new availability system
instead of the old one. This includes making code handle the
following changes from previous commits:

* Removal of availability tables course_modules_availability,
  course_modules_avail_fields, course_sections_availability,
  course_sections_avail_fields.
* Removal of availability fields showavailability,
  availablefrom/until, and (on sections) groupingid.
* Removal of groupmembersonly.

Most of the work here is based on global search for the table,
field, and class names of the old system.

In general the changes in this commit are not intended to alter
behaviour, except in the parts where the old system was removed
from the user interface.