MDL-44723 Modinfo: Allow property access without magic function
While obtaining dynamic data for modinfo, it is not possible to
access even basic properties of cm_info (specifically, name)
because strange PHP behaviour prevents recursive calls to magic
properties. As a result of this, the code will fail if the first
thing you do that requires dynamic data in modinfo is an access
to ->name, but will succeed if the first thing is an access to
anything else. This can cause it to fail on random pages.

This commit adds a new API function that lets you access the
property in exactly the same way as the magic function. Because
it is not a magic function, recursive calls now work in the
expected manner.