MDL-67211 Tasks: Option to disable all background tasks
A new admin setting task_disable_processing (on the existing
settings page for tasks) turns off all task running so that no
new tasks will be started. This is intended for use prior to
an upgrade so that we can safely restart servers that might
otherwise be running tasks.

If turned on, a warning about the new setting displays on the
scheduled tasks screen.

The CLI scripts for running tasks will normally honour this
setting, but you can use --force to run a task even when the
setting is disabled.